Move initialization to _init
